# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/admin/accounts/{id}":{"get":{"tags":["admin-controller"],"operationId":"getAccountById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}},"put":{"tags":["admin-controller"],"operationId":"updateAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}},"delete":{"tags":["admin-controller"],"operationId":"deleteAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/user/transaction":{"get":{"tags":["transaction-controller"],"operationId":"getTransactions","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTransactionResponse"}}}}}},"post":{"tags":["transaction-controller"],"operationId":"createTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","parameters":[{"name":"refresh_token","in":"cookie","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","parameters":[{"name":"refresh_token","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/admin/accounts":{"get":{"tags":["admin-controller"],"operationId":"getAllAccounts","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAccountResponse"}}}}}},"post":{"tags":["admin-controller"],"operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/api/v1/user/me":{"get":{"tags":["user-controller"],"operationId":"getMyProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/api/v1/user/balance":{"get":{"tags":["balance-controller"],"operationId":"getBalance","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BalanceResponse"}}}}}}}},"components":{"schemas":{"UpdateUserRequest":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"}}},"AccountResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"role":{"type":"string","enum":["ADMIN","USER"]}}},"CreateTransactionRequest":{"type":"object","properties":{"description":{"type":"string","minLength":1},"amount":{"type":"number","minimum":0.01},"indicator":{"type":"string","enum":["DB","CR"]}},"required":["amount","description","indicator"]},"TransactionResponse":{"type":"object","properties":{"date":{"type":"string","format":"date-time"},"description":{"type":"string"},"amount":{"type":"number"},"indicator":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6},"address":{"type":"string"}},"required":["email","name","password"]},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"userId":{"type":"integer","format":"int64"},"email":{"type":"string"},"role":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"CreateUserRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"email":{"type":"string","format":"email","maxLength":150,"minLength":0},"password":{"type":"string","maxLength":100,"minLength":6},"address":{"type":"string","maxLength":250,"minLength":0}},"required":["address","email","name","password"]},"PageTransactionResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"BalanceResponse":{"type":"object","properties":{"date":{"type":"string","format":"date-time"},"amount":{"type":"number"},"indicator":{"type":"string"}}},"PageAccountResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}}}}}